chore: move test/ inside src/ to avoid rootDir errors
