+++ packages/sidetree.js/src/test/__fixtures__/didActorDidDocument.json
