+++ packages/sidetree.js/src/test/__fixtures__/simpleDidDocument.json
