+++ packages/sidetree.js/src/test/create.test.ts
@@ -2,9 +2,9 @@ import {
   generateCommitRevealPair,
   generateCreateOperationRequest,
 } from './operations_helper';
-import DocumentComposer from '../src/DocumentComposer';
-import CreateOperation from '../src/CreateOperation';
-import Jwk from '../src/util/Jwk';
+import DocumentComposer from '../DocumentComposer';
+import CreateOperation from '../CreateOperation';
+import Jwk from '../util/Jwk';
 import JwkEs256k from 'models/JwkEs256k';
 import didActorDidDocument from './__fixtures__/didActorDidDocument.json';
 import simpleDidDocument from './__fixtures__/simpleDidDocument.json';
